Disable transaction checks that depend on a broken API
The transaction verifier and its sub-verifiers need to validate
transactions based on the block's network upgrade.

To find the network upgrade, we need the block height and network. But
the transaction verifier requests and data structures don't provide this
information.

So we temporarily disable all the transaction verifier checks that depend
on a hard-coded network upgrade.

Temporary workaround for #1367.